+++ helltfbot.ts
@@ -1,9 +1,8 @@
 import 'dotenv/config'
 import { TwitchBot } from './src/client/bot.js'
 import { mainClient } from './src/client/main-bot.js'
-import { watchClient } from './src/client/track-bot.js'
 
-globalThis.hb = new TwitchBot(mainClient, watchClient)
+globalThis.hb = new TwitchBot(mainClient)
 ;(async () => {
   await hb.init()
   hb.joinChannels()